Multithreading Memory Library task A library, that maintains work with allocating memory and giving it to threads. Also there is a testing target file.
There are class and struct:

-Memory pool class;

-Occupied inerval struct;

Memory pool class has pointer to it's memory begining, size of
allocated memory and a list of occupied intervals, any of which has
pointer to begining of it's allocated memory and pointer to end of it.

MemoryPool class works as allocator for user, and has methods, that
return pointer to allocated/reallocated memory of given length.
Base constructor gives memory pool for 8 megabytes.

Testing file has function (UserThread) that is performed on threads
concurrently. Each thread asking for allocating memory and fills it
with it's own unique value, then thread asking memory pool to
reallocate memory and if it got bigger, fill the rest with that
unique number.

Signed-off-by: Oleksandr Volkov <oleksandr.v.volkov@globallogic.com>